refactor: 최단 경로 찾을 떄 GraphService에서 찾도록 수정
