+++ src/main/java/wooteco/subway/admin/service/LineService.java
@@ -3,32 +3,26 @@ package wooteco.subway.admin.service;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.jgrapht.GraphPath;
 import org.springframework.stereotype.Service;
 
-import wooteco.subway.admin.domain.Edge;
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.domain.LineStation;
-import wooteco.subway.admin.domain.PathType;
 import wooteco.subway.admin.domain.Station;
 import wooteco.subway.admin.dto.LineDetailResponse;
 import wooteco.subway.admin.dto.LineRequest;
 import wooteco.subway.admin.dto.LineStationCreateRequest;
-import wooteco.subway.admin.dto.PathResponse;
 import wooteco.subway.admin.dto.WholeSubwayResponse;
 import wooteco.subway.admin.repository.LineRepository;
 import wooteco.subway.admin.repository.StationRepository;
 
 @Service
 public class LineService {
-	private LineRepository lineRepository;
-	private StationRepository stationRepository;
-	private GraphService graphService;
+	private final LineRepository lineRepository;
+	private final StationRepository stationRepository;
 
-	public LineService(LineRepository lineRepository, StationRepository stationRepository, GraphService graphService) {
+	public LineService(LineRepository lineRepository, StationRepository stationRepository) {
 		this.lineRepository = lineRepository;
 		this.stationRepository = stationRepository;
-		this.graphService = graphService;
 	}
 
 	public Line save(Line line) {
@@ -76,22 +70,5 @@ public class LineService {
 		return lines.stream()
 			.map(line -> findLineWithStationsById(line.getId()))
 			.collect(Collectors.collectingAndThen(Collectors.toList(), WholeSubwayResponse::of));
-	}
-
-	public PathResponse searchPath(String source, String target, PathType pathType) {
-		List<Line> lines = lineRepository.findAll();
-		List<Station> stations = stationRepository.findAll();
-		Station sourceStation = stationRepository.findByName(source).orElseThrow(RuntimeException::new);
-		Station targetStation = stationRepository.findByName(target).orElseThrow(RuntimeException::new);
-
-		GraphPath<Station, Edge> path = graphService.findPath(lines, stations, sourceStation, targetStation, pathType);
-
-		int distance = path.getEdgeList().stream()
-			.mapToInt(Edge::getDistance)
-			.sum();
-		int duration = path.getEdgeList().stream()
-			.mapToInt(Edge::getDuration)
-			.sum();
-		return PathResponse.of(path.getVertexList(), distance, duration);
 	}
 }
