+++ src/test/java/wooteco/subway/admin/service/GraphServiceTest.java
@@ -1,80 +1,79 @@
 package wooteco.subway.admin.service;
 
 import static org.assertj.core.api.Assertions.*;
+import static org.mockito.Mockito.*;
 
 import java.time.LocalTime;
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Objects;
+import java.util.Optional;
 
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.graph.WeightedMultigraph;
+import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.jdbc.Sql;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 
-import wooteco.subway.admin.domain.Edge;
 import wooteco.subway.admin.domain.Line;
 import wooteco.subway.admin.domain.LineStation;
 import wooteco.subway.admin.domain.PathType;
+import wooteco.subway.admin.domain.Station;
+import wooteco.subway.admin.dto.PathResponse;
+import wooteco.subway.admin.repository.LineRepository;
+import wooteco.subway.admin.repository.StationRepository;
 
 /**
  *    class description
  *
  *    @author HyungJu An, KuenHwi Choi
  */
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Sql("/truncate.sql")
+@ExtendWith(MockitoExtension.class)
 class GraphServiceTest {
+	@Mock
+	private LineRepository lineRepository;
+	@Mock
+	private StationRepository stationRepository;
+
+	private GraphService graphService;
+
 	private Line line1;
 	private Line line2;
 
-	@Autowired
-	private GraphService graphService;
-
 	@BeforeEach
 	void setUp() {
-		line1 = Line.of("2호선", LocalTime.of(05, 30), LocalTime.of(22, 30), 5).withId(1L);
-		line2 = Line.of("신분당선", LocalTime.of(5, 30), LocalTime.of(22, 30), 5).withId(2L);
-	}
+		graphService = new GraphService(lineRepository, stationRepository);
 
-	@DisplayName("유효한 출발역과 도착역이 주어지면 적절한 경로를 반환하는지 테스트한다.")
-	@Test
-	void findPath() {
+		line1 = Line.of("2호선", LocalTime.of(5, 30), LocalTime.of(22, 30), 5).withId(1L);
 		line1.addLineStation(new LineStation(null, 1L, 10, 10));
 		line1.addLineStation(new LineStation(1L, 2L, 10, 10));
 		line1.addLineStation(new LineStation(2L, 3L, 10, 10));
 
+		line2 = Line.of("신분당선", LocalTime.of(5, 30), LocalTime.of(22, 30), 5).withId(2L);
 		line2.addLineStation(new LineStation(null, 6L, 10, 10));
 		line2.addLineStation(new LineStation(6L, 5L, 10, 10));
 		line2.addLineStation(new LineStation(5L, 4L, 10, 10));
 		line2.addLineStation(new LineStation(4L, 1L, 10, 10));
+	}
 
-		WeightedMultigraph<Long, Edge> graph
-			= new WeightedMultigraph(Edge.class);
+	@DisplayName("최단 거리 경로와 최소 시간 경로를 찾는다.")
+	@Test
+	void searchPaths() {
+		when(lineRepository.findAll()).thenReturn(Arrays.asList(this.line1, this.line2));
+		List<Station> stations = Lists.newArrayList(new Station(6L, "청계산입구역"),
+			new Station(5L, "양재시민의숲역"),
+			new Station(4L, "양재역"), new Station(1L, "강남역"),
+			new Station(2L, "역삼역"), new Station(3L, "삼성역")
+		);
+		when(stationRepository.findAll()).thenReturn(stations);
+		when(stationRepository.findByName("청계산입구역")).thenReturn(Optional.of(new Station(6L, "청계산입구역")));
+		when(stationRepository.findByName("삼성역")).thenReturn(Optional.of(new Station(3L, "삼성역")));
 
-		List<Line> lines = new ArrayList<>();
-		lines.add(line1);
-		lines.add(line2);
-		PathType type = PathType.DISTANCE;
+		PathResponse pathResponse = graphService.searchPath("청계산입구역", "삼성역", PathType.DISTANCE);
 
-		lines.stream()
-			.flatMap(it -> it.getLineStationsId().stream())
-			.forEach(graph::addVertex);
-
-		lines.stream()
-			.flatMap(it -> it.getStations().stream())
-			.filter(it -> Objects.nonNull(it.getPreStationId()))
-			.forEach(it -> {
-				Edge edge = Edge.of(it);
-				graph.addEdge(it.getPreStationId(), it.getStationId(), edge);
-				graph.setEdgeWeight(edge, type.findWeightOf(it));
-			});
-
-		DijkstraShortestPath dijkstraShortestPath = new DijkstraShortestPath(graph);
-		assertThat(dijkstraShortestPath.getPath(6L, 3L).getEdgeList().size()).isEqualTo(5);
+		assertThat(pathResponse.getStations().size()).isEqualTo(6);
+		assertThat(pathResponse.getDistance()).isEqualTo(50);
+		assertThat(pathResponse.getDuration()).isEqualTo(50);
 	}
 }
